[MOB-11797] Bootstrap Example (#339)
General refactoring prior to the upcoming PR revamping the example app.
