+++ example/android/.gitignore
