+++ example/android/app/src/androidTest/java/com/instabug/flutter/example/FeatureRequestsUITest.java
@@ -1,11 +1,12 @@
-package com.example.InstabugSample;
+package com.instabug.flutter.example;
 
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.flutter.EspressoFlutter.onFlutterWidget;
-
-import static com.example.InstabugSample.util.InstabugAssertions.assertViewWillBeVisible;
 
 import androidx.test.espresso.flutter.action.FlutterActions;
 import androidx.test.espresso.flutter.matcher.FlutterMatchers;
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.rule.ActivityTestRule;
 
@@ -15,16 +16,17 @@ import org.junit.runner.RunWith;
 
 
 @RunWith(AndroidJUnit4.class)
-public class SurveysUITest {
+public class FeatureRequestsUITest {
     @Rule
     public ActivityTestRule<MainActivity> mActivityRule =
             new ActivityTestRule<>(MainActivity.class);
 
     @Test
-    public void showManualSurvey() {
-        onFlutterWidget(FlutterMatchers.withText("Show Manual Survey"))
+    public void showFeatureRequestsScreen() {
+        onFlutterWidget(FlutterMatchers.withText("Show Feature Requests"))
                 .perform(FlutterActions.scrollTo(), FlutterActions.click());
 
-        assertViewWillBeVisible("instabug_survey_dialog_container", 2000);
+        onView(ViewMatchers.withText("Feature Requests"))
+                .check(matches(ViewMatchers.isDisplayed()));
     }
 }
