+++ example/android/app/src/androidTest/java/com/instabug/flutter/example/SurveysUITest.java
@@ -1,12 +1,11 @@
-package com.example.InstabugSample;
+package com.instabug.flutter.example;
 
-import static androidx.test.espresso.Espresso.onView;
-import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.flutter.EspressoFlutter.onFlutterWidget;
+
+import static com.instabug.flutter.example.util.InstabugAssertions.assertViewWillBeVisible;
 
 import androidx.test.espresso.flutter.action.FlutterActions;
 import androidx.test.espresso.flutter.matcher.FlutterMatchers;
-import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.rule.ActivityTestRule;
 
@@ -16,17 +15,16 @@ import org.junit.runner.RunWith;
 
 
 @RunWith(AndroidJUnit4.class)
-public class FeatureRequestsUITest {
+public class SurveysUITest {
     @Rule
     public ActivityTestRule<MainActivity> mActivityRule =
             new ActivityTestRule<>(MainActivity.class);
 
     @Test
-    public void showFeatureRequestsScreen() {
-        onFlutterWidget(FlutterMatchers.withText("Show Feature Requests"))
+    public void showManualSurvey() {
+        onFlutterWidget(FlutterMatchers.withText("Show Manual Survey"))
                 .perform(FlutterActions.scrollTo(), FlutterActions.click());
 
-        onView(ViewMatchers.withText("Feature Requests"))
-                .check(matches(ViewMatchers.isDisplayed()));
+        assertViewWillBeVisible("instabug_survey_dialog_container", 2000);
     }
 }
