+++ example/android/app/src/androidTest/java/com/instabug/flutter/example/util/HasBackgroundColorMatcher.java
@@ -1,4 +1,4 @@
-package com.example.InstabugSample.util;
+package com.instabug.flutter.example.util;
 
 import android.graphics.Bitmap;
 import android.graphics.Canvas;
