+++ example/android/app/src/main/kotlin/com/example/InstabugSample/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.InstabugSample
+package com.instabug.flutter.example
 
 import io.flutter.embedding.android.FlutterActivity
