+++ example/android/settings.gradle
@@ -1,3 +1,4 @@
+rootProject.name = 'InstabugExample'
 include ':app'
 
 def localPropertiesFile = new File(rootProject.projectDir, "local.properties")
