+++ example/ios/InstabugTests/ArgsRegistryTests.m
