+++ example/ios/InstabugTests/Info.plist
