+++ example/ios/InstabugUITests/Info.plist
