+++ example/ios/Runner.xcodeproj/xcshareddata/xcschemes/Runner.xcscheme
@@ -42,8 +42,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "C090017825D9A030006F3DAE"
-               BuildableName = "InstabugSampleTests.xctest"
-               BlueprintName = "InstabugSampleTests"
+               BuildableName = "InstabugTests.xctest"
+               BlueprintName = "InstabugTests"
                ReferencedContainer = "container:Runner.xcodeproj">
             </BuildableReference>
          </TestableReference>
@@ -52,8 +52,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "C09001B025D9A3C5006F3DAE"
-               BuildableName = "InstabugSampleUITests.xctest"
-               BlueprintName = "InstabugSampleUITests"
+               BuildableName = "InstabugUITests.xctest"
+               BlueprintName = "InstabugUITests"
                ReferencedContainer = "container:Runner.xcodeproj">
             </BuildableReference>
          </TestableReference>
