+++ example/test_driver/example.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_driver/driver_extension.dart';
-import 'package:instabug_sample/main.dart' as app;
+import 'package:instabug_flutter_example/main.dart' as app;
 
 void main() {
   enableFlutterDriverExtension();
